Code improvements.

+++ DevilDaggersAssetCore/ChunkInfo.cs
@@ -5,11 +5,11 @@ namespace DevilDaggersAssetCore
 	public class ChunkInfo
 	{
 		public Type Type { get; set; }
-		public byte[] BinaryTypes { get; set; }
+		public ushort[] BinaryTypes { get; set; }
 		public string FileExtension { get; set; }
 		public string FolderName { get; set; }
 
-		public ChunkInfo(Type type, byte[] binaryTypes, string fileExtension, string folderName)
+		public ChunkInfo(Type type, ushort[] binaryTypes, string fileExtension, string folderName)
 		{
 			Type = type;
 			BinaryTypes = binaryTypes;
